Reorganize hab command module structure
This moves the submodules for each command into their own file. The
100 character line limitation of our auto-formatter and nested
modules don't play well together. Files are free, let's use them!
